Move RiskAnalysisTemplate models to separated files
