+++ packages/models/src/purpose-template/purposeTemplate.ts
@@ -5,8 +5,8 @@ import {
   PurposeTemplateId,
   TenantId,
 } from "../brandedIds.js";
-import { RiskAnalysisFormTemplate } from "../risk-analysis/riskAnalysisTemplate.js";
 import { TenantKind } from "../tenant/tenant.js";
+import { RiskAnalysisFormTemplate } from "../risk-analysis/riskAnalysisTemplate.js";
 
 export const purposeTemplateState = {
   draft: "Draft",
