+++ packages/models/src/risk-analysis/riskAnalysisTemplate.ts
@@ -5,8 +5,8 @@ import {
   RiskAnalysisFormTemplateId,
 } from "../brandedIds.js";
 import {
-  RiskAnalysisSingleAnswer,
   RiskAnalysisMultiAnswer,
+  RiskAnalysisSingleAnswer,
 } from "./riskAnalysis.js";
 
 export const RiskAnalysisTemplateAnswerAnnotationDocument = z.object({
@@ -30,23 +30,24 @@ export type RiskAnalysisTemplateAnswerAnnotation = z.infer<
   typeof RiskAnalysisTemplateAnswerAnnotation
 >;
 
-export const RiskAnalysisTemplateSingleAnswer = RiskAnalysisSingleAnswer.and(
-  z.object({
+export const RiskAnalysisTemplateSingleAnswer = RiskAnalysisSingleAnswer.extend(
+  {
     editable: z.boolean(),
     annotation: RiskAnalysisTemplateAnswerAnnotation.optional(),
+    assistiveText: z.string().optional(),
     suggestedValues: z.array(z.string()),
-  })
+  }
 );
 export type RiskAnalysisTemplateSingleAnswer = z.infer<
   typeof RiskAnalysisTemplateSingleAnswer
 >;
 
-export const RiskAnalysisTemplateMultiAnswer = RiskAnalysisMultiAnswer.and(
-  z.object({
-    editable: z.boolean(),
-    annotation: RiskAnalysisTemplateAnswerAnnotation.optional(),
-  })
-);
+export const RiskAnalysisTemplateMultiAnswer = RiskAnalysisMultiAnswer.extend({
+  editable: z.boolean(),
+  annotation: RiskAnalysisTemplateAnswerAnnotation.optional(),
+  assistiveText: z.string().optional(),
+});
+
 export type RiskAnalysisTemplateMultiAnswer = z.infer<
   typeof RiskAnalysisTemplateMultiAnswer
 >;
