+++ packages/purpose-template-process/src/services/validators.ts
@@ -1,8 +1,9 @@
-import { riskAnalysisValidatedFormTemplateToNewRiskAnalysisFormTemplate } from "pagopa-interop-commons";
-
 import { RiskAnalysisFormTemplate, TenantKind } from "pagopa-interop-models";
 import { purposeTemplateApi } from "pagopa-interop-api-clients";
-import { RiskAnalysisTemplateValidatedForm } from "../model/riskAnalysisFormTemplate.js";
+import {
+  RiskAnalysisTemplateValidatedForm,
+  riskAnalysisValidatedFormTemplateToNewRiskAnalysisFormTemplate,
+} from "../model/riskAnalysisFormTemplate.js";
 import { validatePurposeTemplateRiskAnalysis } from "../model/riskAnalysisTemplateValidation.js";
 import {
   missingFreeOfChargeReason,
