+++ packages/purpose-template-process/test/integration/riskAnalysisTemplateValidation.test.ts
@@ -1,15 +1,12 @@
 import { describe, expect, it } from "vitest";
-
-import {
-  riskAnalysisFormTemplateToRiskAnalysisFormTemplateToValidate,
-  validatePurposeTemplateRiskAnalysis,
-} from "pagopa-interop-commons";
 import { tenantKind } from "pagopa-interop-models";
 import {
   getMockValidRiskAnalysisFormTemplate,
   validatedRiskAnalysisTemplate2_0_Private,
   validatedRiskAnalysisTemplate3_0_Pa,
 } from "../riskAnalysisTestUtils.js";
+import { riskAnalysisFormTemplateToRiskAnalysisFormTemplateToValidate } from "../../src/model/riskAnalysisFormTemplate.js";
+import { validatePurposeTemplateRiskAnalysis } from "../../src/model/riskAnalysisTemplateValidation.js";
 
 describe("Risk Analysis Template Validation", () => {
   it("should succeed on correct form 3.0 on tenant kind PA", () => {
